+++ example/react-example/.ddotrc.ts
@@ -13,6 +13,11 @@ export default {
         alias: {
           config: join(__dirname, './src/config'),
         },
+        tsLoaderOption: {
+          transformers: {
+            before: [require('ts-react-hot-transformer')],
+          },
+        },
         define: {
           c: { a: 'c' },
         },
@@ -21,7 +26,6 @@ export default {
             .entry('index')
             .add('./src/index')
             .end();
-
         },
       },
     ],