+++ example/vue-example/src/pages/App.vue
@@ -2,7 +2,7 @@
   <div class="example">{{ msg }}</div>
 </template>
 
-<script>
+<script lang="ts">
 export default {
   data () {
     return {