+++ packages/plugin-webpack/src/config/env/dev.ts
@@ -6,5 +6,4 @@ export default (config: Config, cfgset: ICFG,{path }) => {
   config.plugin('mini-css').use(require('mini-css-extract-plugin'));
   config.plugin('hot-module').use(webpack.HotModuleReplacementPlugin);
   config.devtool('cheap-module-source-map');
-
 };