+++ example/vue-example/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.10](https://github.com/Jetsly/ddot/compare/vue@1.0.9...vue@1.0.10) (2019-01-28)
+
+**Note:** Version bump only for package vue
+
+
+
+
+
+## [1.0.9](https://github.com/Jetsly/ddot/compare/vue@1.0.8...vue@1.0.9) (2019-01-27)
+
+**Note:** Version bump only for package vue
+
+
+
+
+
+## [1.0.8](https://github.com/Jetsly/ddot/compare/vue@1.0.7...vue@1.0.8) (2019-01-22)
+
+**Note:** Version bump only for package vue
+
+
+
+
+
+## [1.0.7](https://github.com/Jetsly/ddot/compare/vue@1.0.6...vue@1.0.7) (2019-01-22)
+
+**Note:** Version bump only for package vue
+
+
+
+
+
+## [1.0.6](https://github.com/Jetsly/ddot/compare/vue@1.0.5...vue@1.0.6) (2019-01-14)
+
+**Note:** Version bump only for package vue
+
+
+
+
+
+## [1.0.5](https://github.com/Jetsly/ddot/compare/vue@1.0.4...vue@1.0.5) (2019-01-14)
+
+
+### Features
+
+* **ddot-plugin-webpack:** change fastify=>express ([8d6e780](https://github.com/Jetsly/ddot/commit/8d6e780))
+
+
+
+
+
+## [1.0.4](https://github.com/Jetsly/ddot/compare/vue@1.0.3...vue@1.0.4) (2019-01-11)
+
+**Note:** Version bump only for package vue
+
+
+
+
+
 ## [1.0.3](https://github.com/Jetsly/ddot/compare/vue@1.0.2...vue@1.0.3) (2019-01-08)
 
 **Note:** Version bump only for package vue