+++ packages/plugin-utils/CHANGELOG.md
@@ -3,6 +3,39 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.7](https://github.com/Jetsly/ddot/compare/@ddot/plugin-utils@0.0.6...@ddot/plugin-utils@0.0.7) (2019-01-27)
+
+
+### Bug Fixes
+
+* **ddot-plugin-webpack:** fix mjs ([fb52339](https://github.com/Jetsly/ddot/commit/fb52339))
+
+
+
+
+
+## [0.0.6](https://github.com/Jetsly/ddot/compare/@ddot/plugin-utils@0.0.5...@ddot/plugin-utils@0.0.6) (2019-01-22)
+
+
+### Bug Fixes
+
+* **ddot-plugin-webpack:** fix publish ([8793994](https://github.com/Jetsly/ddot/commit/8793994))
+
+
+
+
+
+## [0.0.5](https://github.com/Jetsly/ddot/compare/@ddot/plugin-utils@0.0.4...@ddot/plugin-utils@0.0.5) (2019-01-22)
+
+
+### Features
+
+* **ddot-plugin-webpack:** add sourcemap when prod ([45b63b4](https://github.com/Jetsly/ddot/commit/45b63b4))
+
+
+
+
+
 ## [0.0.4](https://github.com/Jetsly/ddot/compare/@ddot/plugin-utils@0.0.3...@ddot/plugin-utils@0.0.4) (2019-01-08)
 
 **Note:** Version bump only for package @ddot/plugin-utils